api/requirements.txt
tensorflow>=2.13.0
numpy>=1.22.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=1.5.0
Pillow>=9.5.0
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6

